Bug fixed: like icons wasn't updating on time

+++ shared/lib/likes/index.ts
@@ -6,22 +6,20 @@ class Like extends InitService {
       .from("Like")
       .insert({ video_id, user_id, video_user_id: user_id })
       .select()
-      .maybeSingle();
+      .limit(1);
 
     if (error) throw error;
     return data;
   }
 
-  async removeLikeVideo(like_id: string) {
-    const { data, error } = await this.client
+  async removeLikeVideo(video_id: string, user_id: string) {
+    const { error } = await this.client
       .from("Like")
       .delete()
-      .eq("id", like_id)
-      .select()
-      .maybeSingle();
+      .eq("user_id", user_id)
+      .eq("video_id", video_id);
 
     if (error) throw error;
-    return data;
   }
 
   async getLikesByUser(user_id: string) {
